requirements.txt
seqeval~=1.2.2
tqdm~=4.49.0
torchmetrics~=0.3.2
numpy~=1.20.2
keras~=2.4.3
attrs~=21.2.0
tensorflow~=2.3.0
tokenizers~=0.10.2
six~=1.15.0
sklearn~=0.0
scikit-learn~=0.22
scipy~=1.6.2
filelock~=3.0.12
packaging~=20.9
ipython~=7.22.0
importlib_metadata~=3.10.0
sentencepiece~=0.1.92
pillow~=8.2.0
requests~=2.25.1
torchaudio~=0.8.1
tensorboard~=2.4.1
